requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohttp==3.9.5
aiosignal==1.3.1
altgraph==0.17.4
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.37.25
botocore==1.37.25
buildozer==1.5.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.43.0
cloudipsp==1.0.4
colorama==0.4.6
comtypes==1.4.8
coverage==7.6.4
CurrencyConverter==0.17.33
distlib==0.3.9
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1.5
django-allauth==65.5.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-extensions==3.2.3
django-heroku==0.3.1
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docutils==0.21.2
filelock==3.16.1
Flask==3.1.0
flask-cors==5.0.1
Flask-RESTful==0.3.10
frozenlist==1.4.1
google-api-core==2.22.0
google-auth==2.36.0
google-cloud-speech==2.28.0
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
gTTS==2.5.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-examples==2.3.0
Kivy-Garden==0.1.5
kivymd==1.2.0
magic-filter==1.0.12
MarkupSafe==3.0.2
MouseInfo==0.1.3
multidict==6.0.5
mysql-connector-python==9.1.0
numexpr==2.10.1
numpy==2.1.0
packaging==24.1
pefile==2023.2.7
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.28.3
psycopg2==2.9.10
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.8.2
pydantic_core==2.20.1
PyGetWindow==0.0.9
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.9.0
PyMsgBox==1.0.9
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
pyTelegramBotAPI==4.23.0
pytesseract==0.3.13
pytest==8.3.3
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.7
python-weather==2.0.3
pyttsx3==2.98
pyttsx4==3.0.15
pytweening==1.2.0
pytz==2025.1
pywin32==308
pywin32-ctypes==0.2.3
requests==2.32.3
rsa==4.9
s3transfer==0.11.4
setuptools==73.0.0
sh==2.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.11.0
sqlparse==0.5.3
tesseract==0.1.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
virtualenv==20.27.1
watchdog==5.0.3
Werkzeug==3.1.3
whitenoise==6.9.0
yarl==1.9.4
gunicorn==23.0.0
psycopg2-binary==2.9.9
